backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pysam>=0.22.0
httpx>=0.27.0
redis>=5.0.0
reportlab>=4.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
passlib>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
